--------------------------           NOTE --------------------------
There was a major difference between my local-Windows (Babun/Cygwin) and local-Linux, and there may be a nightmare of stuff hidden in the previous revision(s).  Who knows.. but things look good.  Fingers crossed that git has archived anything that might have gone awry and may need to be recovered.